// my-movie-app/src/model/movies_coming.ts
export default interface movies_coming {
    id: string; 
    title: string; 
    year: string; 
    genres: string[]; 
    ratings: number[]; 
    poster: string; 
    contentRating: string; 
    duration: string; 
    releaseDate: string; 
    averageRating: number; 
    originalTitle: string; 
    storyline: string; 
    actors: string[]; 
    imdbRating: any; 
    posterurl: string;

}